+++ custom_components/soundtouchplus/const.py
@@ -2,12 +2,17 @@
 Constants for the Bose SoundTouchPlus component.
 """
 DOMAIN = "soundtouchplus"
+""" Domain identifier for this integration. """
+
+DOMAIN_SPOTIFYPLUS:str = "spotifyplus"
+""" Domain identifier of the SpotifyPlus integration. """
 
 CONF_DEVICE_NAME = "device_name"
 CONF_DEVICE_ID = "device_id"
 CONF_PING_WEBSOCKET_INTERVAL = "ping_websocket_interval"
 CONF_PORT_WEBSOCKET = "port_websocket"
 CONF_OPTION_SOURCE_LIST = "source_list"
+CONF_OPTION_SPOTIFY_MEDIAPLAYER_ENTITY_ID = "spotify_mediaplayer_entity_id"
 
 DEFAULT_PING_WEBSOCKET_INTERVAL = 0
 DEFAULT_PORT = 8090